refactor web feature organization
