+++ .github/workflows/preview.yml
@@ -41,7 +41,7 @@ jobs:
               packages/api-reference/*|\
               packages/mixedbread/*|\
               apps/web/src/features/api-reference/*|\
-              apps/web/src/services/search/domain.ts|\
+              apps/web/src/features/search/domain.ts|\
               apps/web/package.json|\
               packages/mixedbread/package.json|\
               package.json|\
