+++ apps/web/src/features/search/analytics.ts
