+++ apps/web/src/features/search/preferences.ts
